fix(v3/windows): fix IDropTarget callback panic on 32-bit
syscall.NewCallback rejects any parameter wider than a uintptr. The
IDropTarget thunks take POINTL by value (8 bytes), which is one word on
64-bit but two on GOARCH=386. Because the vtable is registered from an
unconditional init() in pkg/w32, every app panicked at startup before
main, whether or not it used drag-and-drop.

Split the three affected thunks into architecture-specific files. On 386
they receive the POINTL as the two 4-byte stack slots stdcall pushes
and reassemble it, mirroring putrect_386.go.
